Refactored projects.php.
Checked our REQUEST_METHOD first, then checked we were engaging in
permissible actions.
Created functions that use our global-scope action lists and ProjectDB
link.
Added a simple switch function that will take care of our requests with
an error if we manage to pass a value that sneaks past our array and is
not part of our program.